requirements.txt
Flask==3.0.2
pymongo==4.6.2
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
langdetect==1.0.9
transformers==4.38.2
torch==2.2.1
certifi==2024.2.2
gunicorn==21.2.0
